+++ packages/storybook/stories/ElementsStories/Dialog.stories.tsx
@@ -1,8 +1,15 @@
-import { useState } from "react";
+import { useRef, useState } from "react";
 
 import { Story } from "@storybook/blocks";
 import type { Meta, StoryObj } from "@storybook/react";
 
+import {
+  Combobox,
+  Tabs,
+  TabsContent,
+  TabsList,
+  TabsTrigger
+} from "@sikka/hawa/elements";
 import { Button } from "@sikka/hawa/elements/button";
 import {
   DialogHeader,
@@ -346,6 +353,79 @@ export const MultistepCarousel: Story = {
               </DialogStep>
             </DialogCarousel>
           </DialogCarouselContent>
+        </Dialog>
+      </div>
+    );
+  }
+};
+export const Playground: Story = {
+  render: (args: any, globals: any) => {
+    const locale = globals.globals?.locale === "ar" ? "ar" : "en";
+    const direction = locale === "ar" ? "rtl" : "ltr";
+    setLocale(locale);
+
+    const [openDialog, setOpenDialog] = useState(true);
+    return (
+      <div className="hawa-flex hawa-flex-row hawa-gap-2">
+        <Button onClick={() => setOpenDialog(!openDialog)}>Open Dialog</Button>
+
+        <Dialog onOpenChange={setOpenDialog} open={openDialog}>
+          <DialogContent>
+            <Combobox
+              data={[
+                { label: "Option 1", value: "1" },
+                { label: "Option 2", value: "2" },
+                { label: "Option 3", value: "3" },
+                { label: "Option 4", value: "4" },
+                { label: "Option 5", value: "5" },
+                { label: "Option 6", value: "6" },
+                { label: "Option 7", value: "7" },
+                { label: "Option 8", value: "8" },
+                { label: "Option 9", value: "9" },
+                { label: "Option 10", value: "10" },
+                { label: "Option 11", value: "11" },
+                { label: "Option 12", value: "12" },
+                { label: "Option 13", value: "13" },
+                { label: "Option 14", value: "14" },
+                { label: "Option 15", value: "15" },
+                { label: "Option 16", value: "16" },
+                { label: "Option 17", value: "17" },
+                { label: "Option 18", value: "18" },
+                { label: "Option 19", value: "19" },
+                { label: "Option 20", value: "20" },
+                { label: "Option 21", value: "21" },
+                { label: "Option 22", value: "22" },
+                { label: "Option 23", value: "23" },
+                { label: "Option 24", value: "24" },
+                { label: "Option 25", value: "25" },
+                { label: "Option 26", value: "26" },
+                { label: "Option 27", value: "27" },
+                { label: "Option 28", value: "28" },
+                { label: "Option 29", value: "29" },
+                { label: "Option 30", value: "30" },
+                { label: "Option 31", value: "31" },
+                { label: "Option 32", value: "32" },
+                { label: "Option 33", value: "33" },
+                { label: "Option 34", value: "34" },
+                { label: "Option 35", value: "35" },
+                { label: "Option 36", value: "36" },
+                { label: "Option 37", value: "37" },
+                { label: "Option 38", value: "38" },
+                { label: "Option 39", value: "39" },
+                { label: "Option 40", value: "40" },
+                { label: "Option 41", value: "41" },
+                { label: "Option 42", value: "42" },
+                { label: "Option 43", value: "43" },
+                { label: "Option 44", value: "44" },
+                { label: "Option 45", value: "45" },
+                { label: "Option 46", value: "46" },
+                { label: "Option 47", value: "47" },
+                { label: "Option 48", value: "48" },
+                { label: "Option 49", value: "49" },
+                { label: "Option 50", value: "50" }
+              ]}
+            />
+          </DialogContent>
         </Dialog>
       </div>
     );